Add latitude, longitude, and submitter_id to restaurant seeds

+++ backend/db/schema.rb
@@ -19,8 +19,8 @@ ActiveRecord::Schema.define(version: 2021_04_11_020529) do
     t.string "name", null: false
     t.string "location", null: false
     t.integer "submitter_id", null: false
-    t.integer "latitude", null: false
-    t.integer "longitude", null: false
+    t.float "latitude", null: false
+    t.float "longitude", null: false
     t.text "description", null: false
     t.index ["latitude", "longitude"], name: "index_restaurants_on_latitude_and_longitude", unique: true
     t.index ["name"], name: "index_restaurants_on_name"
